src/third_party/blink/web_tests/serial/serialPort_readable_disconnect.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/unguessable_token.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/serial/serial.mojom.js"></script>
<script src="resources/serial-test-utils.js"></script>
<script>

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  // Select a buffer size smaller than the amount of data transferred.
  await port.open({ baudrate: 9600, buffersize: 64 });

  const data = new Uint8Array([1, 2, 3, 4, 5, 6, 7, 8]);
  fakePort.write(data);
  fakePort.simulateDisconnectOnRead();

  const reader = port.readable.getReader();
  const { value, done } = await reader.read();
  assert_false(done);
  compareArrays(data, value);

  await promise_rejects_dom(t, 'NetworkError', reader.read());
  assert_equals(port.readable, null);

  await port.close();
}, 'Disconnect error closes readable and sets it to null');

</script>
